Changed truck to transit, Creater (empty) contianer for trailer,Fixed trailer
